Move jsonStringify test from class.spec.ts (serializable class restoration)
- Moved test from packages/run-types/src/runType/collection/class.spec.ts:87-94
- Added imports for jitUtils and DataOnly
- All tests passing
